Rename toggle scripts file
